Changed weights from ints to doubles

+++ src/main/java/redis/clients/jedis/ZParams.java
@@ -23,9 +23,9 @@ public class ZParams {
 
     private List<byte[]> params = new ArrayList<byte[]>();
 
-    public ZParams weights(final int... weights) {
+    public ZParams weights(final double... weights) {
 	params.add(WEIGHTS.raw);
-	for (final int weight : weights) {
+	for (final double weight : weights) {
 	    params.add(Protocol.toByteArray(weight));
 	}
